+++ src/theme/NavbarItem/ComponentTypes.js
@@ -1,5 +1,5 @@
 import ComponentTypes from '@theme-original/NavbarItem/ComponentTypes';
-import ConnectButton from '@site/src/components/connect-button';
+import ConnectButton from '@site/src/components/ConnectButton';
 
 export default {
   ...ComponentTypes,
